+++ backend/AccountingInventory.Core/DTOs/AccountDtos.cs
@@ -13,6 +13,6 @@ namespace AccountingInventory.Core.DTOs
     {
         public decimal Amount { get; set; }
         public string Description { get; set; } = string.Empty;
-        public int CounterpartAccountId { get; set; }
+        public long CounterpartAccountId { get; set; }
     }
 }
