+++ backend/AccountingInventory.Core/DTOs/ReportParams.cs
@@ -8,9 +8,9 @@ namespace AccountingInventory.Core.DTOs
         public DateTime? EndDate { get; set; }
         public string? Status { get; set; }
         public string? PaymentStatus { get; set; }
-        public int? CustomerId { get; set; }
-        public int? SupplierId { get; set; }
-        public int? AccountId { get; set; }
+        public long? CustomerId { get; set; }
+        public long? SupplierId { get; set; }
+        public long? AccountId { get; set; }
         public string? UserId { get; set; }
     }
 }
