+++ backend/AccountingInventory.Core/Interfaces/IGenericRepository.cs
@@ -6,7 +6,7 @@ namespace AccountingInventory.Core.Interfaces
 {
     public interface IGenericRepository<T> where T : BaseEntity
     {
-        Task<T?> GetByIdAsync(int id);
+        Task<T?> GetByIdAsync(long id);
         Task<IReadOnlyList<T>> GetAllAsync();
         Task<IReadOnlyList<T>> FindAsync(Expression<Func<T, bool>> predicate);
         Task<T> AddAsync(T entity);
