+++ backend/AccountingInventory.Core/Interfaces/IProductRepository.cs
@@ -6,6 +6,6 @@ namespace AccountingInventory.Core.Interfaces
     public interface IProductRepository : IGenericRepository<Product>
     {
         Task<Pagination<Product>> GetProductsWithDetailsAsync(ProductSpecParams specParams);
-        Task<Product?> GetProductWithDetailsAsync(int id);
+        Task<Product?> GetProductWithDetailsAsync(long id);
     }
 }
